refactor: moves SerdeEncryptPublicKey
